+++ i2c_lcd/yaz180/include/newlib/i2c_lcd.h
@@ -492,11 +492,11 @@ __OPROTO(,,uint8_t,,LCD_ReadByteFromReg,enum LCD_RegAddress regAddr)
 //  void LCD_WriteByteToReg(enum LCD_RegAddress regAddr, const uint8_t byte);
 __OPROTO(,,void,,LCD_WriteByteToReg,enum LCD_RegAddress regAddr, const uint8_t byte)
 
-//  void LCD_ReadSeriesFromReg(enum LCD_RegAddress regAddr, uint8_t *buf, uint8_t length);
-__OPROTO(,,void,,LCD_ReadSeriesFromReg,enum LCD_RegAddress regAddr, uint8_t *buf, uint8_t length)
+//  void LCD_ReadSeriesFromReg(enum LCD_RegAddress regAddr, uint8_t *buf, uint16_t length);
+__OPROTO(,,void,,LCD_ReadSeriesFromReg,enum LCD_RegAddress regAddr, uint8_t *buf, uint16_t length)
 
-//  void LCD_WriteSeriesToReg(enum LCD_RegAddress regAddr, const uint8_t *buf, uint8_t length);
-__OPROTO(,,void,,LCD_WriteSeriesToReg,enum LCD_RegAddress regAddr, const uint8_t *buf, uint8_t length)
+//  void LCD_WriteSeriesToReg(enum LCD_RegAddress regAddr, const uint8_t *buf, uint16_t length);
+__OPROTO(,,void,,LCD_WriteSeriesToReg,enum LCD_RegAddress regAddr, const uint8_t *buf, uint16_t length)
 
 #ifdef  SUPPORT_FULL_API_LIB
 #if  SUPPORT_FULL_API_LIB == TRUE
@@ -555,11 +555,11 @@ __OPROTO(,,uint8_t,,LCD_ReadByteDispRAM,uint8_t x, uint8_t y)
 //  void LCD_WriteByteDispRAM(uint8_t byte, uint8_t x, uint8_t y);
 __OPROTO(,,void,,LCD_WriteByteDispRAM,uint8_t byte, uint8_t x, uint8_t y)
 
-//  void LCD_ReadSeriesDispRAM(uint8_t *buf, uint8_t length, uint8_t x, uint8_t y);
-__OPROTO(,,void,,LCD_ReadSeriesDispRAM,uint8_t *buf, uint8_t length, uint8_t x, uint8_t y)
+//  void LCD_ReadSeriesDispRAM(uint8_t *buf, uint16_t length, uint8_t x, uint8_t y);
+__OPROTO(,,void,,LCD_ReadSeriesDispRAM,uint8_t *buf, uint16_t length, uint8_t x, uint8_t y)
 
-//  void LCD_WriteSeriesDispRAM(uint8_t *buf, uint8_t length, uint8_t x, uint8_t y);
-__OPROTO(,,void,,LCD_WriteSeriesDispRAM,uint8_t *buf, uint8_t length, uint8_t x, uint8_t y)
+//  void LCD_WriteSeriesDispRAM(uint8_t *buf, uint16_t length, uint8_t x, uint8_t y);
+__OPROTO(,,void,,LCD_WriteSeriesDispRAM,uint8_t *buf, uint16_t length, uint8_t x, uint8_t y)
 
 //  void LCD_DisplayConf(enum LCD_DisplayMode mode);
 __OPROTO(,,void,,LCD_DisplayConf,enum LCD_DisplayMode mode)
